Add api view with project member editing
