+++ teamwork/views.py
@@ -90,33 +90,3 @@ def login(request):
     }
 
     return render(request, 'login.html', context)
-
-
-@require_POST
-def edit_project_members(request, username, project_name):
-    user = get_object_or_404(User, username=username)
-
-    project_name = unquote_plus(project_name)
-    project = get_object_or_404(Project,
-        name=project_name,
-        owner__user__username=username)
-    
-    # POST['remove[]'] contains a list of usernames that
-    # should be removed from the project
-    for member_name in request.POST.getlist('remove[]'):
-        try:
-            member = User.objects.get(username=member_name)
-            project.members.remove(member)
-        except:
-            continue
-    
-    # POST['add[]'] contains a list of usernames that
-    # should be added to the project
-    for member_name in request.POST.getlist('add[]'):
-        try:
-            member = User.objects.get(username=member_name)
-            project.members.add(member)
-        except:
-            continue
-
-    return JsonResponse(request.POST)
